fix(server): prevent HTTP listener deadlock under queue saturation
Closes #842

Three-part fix for the server becoming unresponsive after the task
queue saturates and clients disconnect mid-wait:

1. Wrap the blocking `task_queue.acquire()` call in `enqueue_task`
   with a 30 s `tokio::time::timeout`. On Elapsed the future is
   dropped, which triggers AcquireGuard::Drop and reclaims the
   permit — preventing orphan futures from parking on the semaphore
   indefinitely after client disconnect.

2. Add a skipped-cancelled counter in `PriorityPermitQueue::release()`.
   When ≥3 consecutive cancelled waiters are skipped the existing
   `drain_cancelled()` helper bulk-evicts stale heap entries, bounding
   heap growth under cancellation storms.

3. Extract the task_routes inline test module into task_routes_tests.rs
   to keep task_routes.rs under the 800-line limit, and add three
   stress tests covering cancellation and saturation scenarios.
