+++ crates/harness-server/src/http/task_routes.rs
@@ -273,14 +273,17 @@ pub(crate) async fn enqueue_task(
         req.depends_on.clear();
     }
 
-    // Acquire concurrency permit before spawning. Blocks if all slots are
-    // occupied; rejects immediately if the waiting queue is full.
-    let permit = state
-        .concurrency
-        .task_queue
-        .acquire(&project_id, req.priority)
-        .await
-        .map_err(|e| EnqueueTaskError::Internal(e.to_string()))?;
+    // Acquire permit; 30 s timeout prevents orphan futures on client disconnect.
+    let permit = tokio::time::timeout(
+        std::time::Duration::from_secs(30),
+        state
+            .concurrency
+            .task_queue
+            .acquire(&project_id, req.priority),
+    )
+    .await
+    .map_err(|_| EnqueueTaskError::Internal("queue acquisition timed out".to_string()))?
+    .map_err(|e| EnqueueTaskError::Internal(e.to_string()))?;
 
     let agent = select_agent(
         &req,
@@ -797,474 +800,5 @@ pub(super) async fn cancel_task(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn batch_request_deserializes_issues_format() {
-        let json = r#"{"issues": [300, 301, 302], "agent": "claude", "max_rounds": 3, "turn_timeout_secs": 600}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(req.issues, Some(vec![300, 301, 302]));
-        assert_eq!(req.agent.as_deref(), Some("claude"));
-        assert_eq!(req.max_rounds, Some(3));
-        assert_eq!(req.turn_timeout_secs, Some(600));
-        assert!(req.tasks.is_none());
-    }
-
-    #[test]
-    fn batch_request_deserializes_tasks_format() {
-        let json = r#"{"tasks": [{"description": "fix bug X", "issue": 300}, {"description": "add feature Y", "issue": 301}]}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        let tasks = req.tasks.unwrap();
-        assert_eq!(tasks.len(), 2);
-        assert_eq!(tasks[0].description.as_deref(), Some("fix bug X"));
-        assert_eq!(tasks[0].issue, Some(300));
-        assert_eq!(tasks[1].description.as_deref(), Some("add feature Y"));
-        assert_eq!(tasks[1].issue, Some(301));
-        assert!(req.issues.is_none());
-    }
-
-    #[test]
-    fn batch_request_deserializes_tasks_without_issue() {
-        let json = r#"{"tasks": [{"description": "refactor module"}]}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        let tasks = req.tasks.unwrap();
-        assert_eq!(tasks[0].description.as_deref(), Some("refactor module"));
-        assert!(tasks[0].issue.is_none());
-    }
-
-    #[test]
-    fn batch_request_empty_issues_list() {
-        let json = r#"{"issues": []}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        let has_issues = req.issues.as_ref().is_some_and(|v| !v.is_empty());
-        assert!(!has_issues);
-    }
-
-    #[test]
-    fn batch_request_neither_issues_nor_tasks() {
-        let json = r#"{"agent": "claude"}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        let has_issues = req.issues.as_ref().is_some_and(|v| !v.is_empty());
-        let has_tasks = req.tasks.as_ref().is_some_and(|v| !v.is_empty());
-        assert!(!has_issues && !has_tasks);
-    }
-
-    #[test]
-    fn batch_request_deserializes_project_field() {
-        let json = r#"{"issues": [1, 2], "project": "/home/user/my-repo"}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        assert_eq!(
-            req.project,
-            Some(std::path::PathBuf::from("/home/user/my-repo"))
-        );
-    }
-
-    #[test]
-    fn batch_request_project_defaults_to_none() {
-        let json = r#"{"issues": [1]}"#;
-        let req: BatchCreateTaskRequest = serde_json::from_str(json).unwrap();
-        assert!(req.project.is_none());
-    }
-
-    #[tokio::test]
-    async fn resolve_project_from_registry_passes_through_none() {
-        let result = resolve_project_from_registry(None, None).await;
-        assert!(result.unwrap().0.is_none());
-    }
-
-    #[tokio::test]
-    async fn resolve_project_from_registry_passes_through_existing_dir() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().to_path_buf();
-        let result = resolve_project_from_registry(None, Some(path.clone())).await;
-        assert_eq!(result.unwrap().0, Some(path));
-    }
-
-    #[tokio::test]
-    async fn resolve_project_from_registry_no_registry_passes_through_nondir() {
-        // When no registry is available, non-dir paths are returned as-is
-        // (downstream canonicalization handles them).
-        let path = std::path::PathBuf::from("/nonexistent/path");
-        let result = resolve_project_from_registry(None, Some(path.clone())).await;
-        assert_eq!(result.unwrap().0, Some(path));
-    }
-
-    #[tokio::test]
-    async fn resolve_project_from_registry_resolves_id() {
-        let dir = tempfile::tempdir().unwrap();
-        let registry = crate::project_registry::ProjectRegistry::open(&dir.path().join("p.db"))
-            .await
-            .unwrap();
-        registry
-            .register(crate::project_registry::Project {
-                id: "my-repo".to_string(),
-                root: std::path::PathBuf::from("/home/user/my-repo"),
-                name: None,
-                max_concurrent: None,
-                default_agent: None,
-                active: true,
-                created_at: "2026-01-01T00:00:00Z".to_string(),
-            })
-            .await
-            .unwrap();
-
-        let result = resolve_project_from_registry(
-            Some(&registry),
-            Some(std::path::PathBuf::from("my-repo")),
-        )
-        .await;
-        let (path, agent) = result.unwrap();
-        assert_eq!(path, Some(std::path::PathBuf::from("/home/user/my-repo")));
-        assert_eq!(agent, None);
-    }
-
-    #[tokio::test]
-    async fn resolve_project_from_registry_returns_default_agent_from_record() {
-        let dir = tempfile::tempdir().unwrap();
-        let registry = crate::project_registry::ProjectRegistry::open(&dir.path().join("p.db"))
-            .await
-            .unwrap();
-        registry
-            .register(crate::project_registry::Project {
-                id: "pinned-repo".to_string(),
-                root: std::path::PathBuf::from("/home/user/pinned-repo"),
-                name: None,
-                max_concurrent: None,
-                default_agent: Some("opus".to_string()),
-                active: true,
-                created_at: "2026-01-01T00:00:00Z".to_string(),
-            })
-            .await
-            .unwrap();
-
-        let result = resolve_project_from_registry(
-            Some(&registry),
-            Some(std::path::PathBuf::from("pinned-repo")),
-        )
-        .await;
-        let (path, agent) = result.unwrap();
-        assert_eq!(
-            path,
-            Some(std::path::PathBuf::from("/home/user/pinned-repo"))
-        );
-        assert_eq!(agent.as_deref(), Some("opus"));
-    }
-
-    #[tokio::test]
-    async fn resolve_project_from_registry_unknown_id_returns_bad_request() {
-        let dir = tempfile::tempdir().unwrap();
-        let registry = crate::project_registry::ProjectRegistry::open(&dir.path().join("p.db"))
-            .await
-            .unwrap();
-
-        let result = resolve_project_from_registry(
-            Some(&registry),
-            Some(std::path::PathBuf::from("unknown-repo")),
-        )
-        .await;
-        assert!(matches!(result, Err(EnqueueTaskError::BadRequest(_))));
-    }
-
-    #[test]
-    fn conflict_groups_empty_refs_are_singletons() {
-        let refs: Vec<Vec<String>> = vec![vec![], vec![], vec![]];
-        let groups = build_conflict_groups(&refs);
-        assert_eq!(groups.len(), 3);
-        for g in &groups {
-            assert_eq!(g.len(), 1);
-        }
-    }
-
-    #[test]
-    fn conflict_groups_two_tasks_share_file() {
-        let refs = vec![
-            vec!["src/auth.rs".to_string()],
-            vec!["src/auth.rs".to_string(), "src/db.rs".to_string()],
-            vec!["src/config.rs".to_string()],
-        ];
-        let groups = build_conflict_groups(&refs);
-        // Tasks 0 and 1 share auth.rs; task 2 is independent.
-        assert_eq!(groups.len(), 2);
-        assert!(
-            groups.iter().any(|g| g.contains(&0) && g.contains(&1)),
-            "tasks 0 and 1 must be in the same conflict group"
-        );
-        assert!(
-            groups.iter().any(|g| g.len() == 1 && g[0] == 2),
-            "task 2 must be a singleton group"
-        );
-    }
-
-    #[test]
-    fn conflict_groups_transitive_overlap() {
-        // A-B share a.rs, B-C share b.rs → all three in same group.
-        let refs = vec![
-            vec!["a.rs".to_string()],
-            vec!["a.rs".to_string(), "b.rs".to_string()],
-            vec!["b.rs".to_string()],
-        ];
-        let groups = build_conflict_groups(&refs);
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0].len(), 3);
-    }
-
-    #[test]
-    fn conflict_groups_no_overlap() {
-        let refs = vec![
-            vec!["a.rs".to_string()],
-            vec!["b.rs".to_string()],
-            vec!["c.rs".to_string()],
-        ];
-        let groups = build_conflict_groups(&refs);
-        assert_eq!(groups.len(), 3);
-        for g in &groups {
-            assert_eq!(g.len(), 1);
-        }
-    }
-
-    #[test]
-    fn conflict_groups_single_task() {
-        let refs = vec![vec!["src/main.rs".to_string()]];
-        let groups = build_conflict_groups(&refs);
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0], vec![0]);
-    }
-
-    // ── select_agent three-tier precedence tests ─────────────────────────────
-
-    use harness_agents::registry::AgentRegistry;
-    use harness_core::{
-        agent::{AgentRequest, AgentResponse, CodeAgent, StreamItem},
-        config::HarnessConfig,
-        types::{Capability, TokenUsage},
-    };
-
-    struct StubAgent {
-        name: &'static str,
-    }
-
-    #[async_trait::async_trait]
-    impl CodeAgent for StubAgent {
-        fn name(&self) -> &str {
-            self.name
-        }
-        fn capabilities(&self) -> Vec<Capability> {
-            vec![]
-        }
-        async fn execute(&self, _req: AgentRequest) -> harness_core::error::Result<AgentResponse> {
-            Ok(AgentResponse {
-                output: String::new(),
-                stderr: String::new(),
-                items: vec![],
-                token_usage: TokenUsage {
-                    input_tokens: 0,
-                    output_tokens: 0,
-                    total_tokens: 0,
-                    cost_usd: 0.0,
-                },
-                model: self.name.to_string(),
-                exit_code: Some(0),
-            })
-        }
-        async fn execute_stream(
-            &self,
-            _req: AgentRequest,
-            _tx: tokio::sync::mpsc::Sender<StreamItem>,
-        ) -> harness_core::error::Result<()> {
-            Ok(())
-        }
-    }
-
-    fn registry_with(default: &str, names: &[&'static str]) -> AgentRegistry {
-        let mut reg = AgentRegistry::new(default);
-        for &n in names {
-            reg.register(n, Arc::new(StubAgent { name: n }));
-        }
-        reg
-    }
-
-    fn req_with_agent(
-        agent: Option<&str>,
-        project: Option<std::path::PathBuf>,
-    ) -> task_runner::CreateTaskRequest {
-        let mut req = task_runner::CreateTaskRequest::default();
-        req.agent = agent.map(str::to_owned);
-        req.project = project;
-        req.prompt = Some("do something".to_string());
-        req
-    }
-
-    #[test]
-    fn agent_override_wins() {
-        let reg = registry_with("auto", &["opus", "claude"]);
-        let mut server = HarnessConfig::default();
-        server.agents.default_agent = "claude".to_string();
-
-        // Build a temp project dir with no .harness/config.toml
-        let dir = tempfile::tempdir().unwrap();
-        let req = req_with_agent(Some("opus"), Some(dir.path().to_path_buf()));
-
-        let agent = select_agent(&req, &reg, None).unwrap();
-        assert_eq!(agent.name(), "opus");
-    }
-
-    #[test]
-    fn project_default_used() {
-        let reg = registry_with("auto", &["claude", "sonnet"]);
-        let mut server = HarnessConfig::default();
-        server.agents.default_agent = "auto".to_string();
-
-        // Write a project config that sets agent.default = "claude"
-        let dir = tempfile::tempdir().unwrap();
-        let harness_dir = dir.path().join(".harness");
-        std::fs::create_dir_all(&harness_dir).unwrap();
-        std::fs::write(
-            harness_dir.join("config.toml"),
-            "[agent]\ndefault = \"claude\"\n",
-        )
-        .unwrap();
-
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-        let agent = select_agent(&req, &reg, None).unwrap();
-        assert_eq!(agent.name(), "claude");
-    }
-
-    #[test]
-    fn global_fallback_when_no_project_config() {
-        let reg = registry_with("sonnet", &["sonnet"]);
-        let mut server = HarnessConfig::default();
-        server.agents.default_agent = "sonnet".to_string();
-
-        // No .harness/config.toml in the temp dir
-        let dir = tempfile::tempdir().unwrap();
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-
-        let agent = select_agent(&req, &reg, None).unwrap();
-        assert_eq!(agent.name(), "sonnet");
-    }
-
-    #[test]
-    fn global_fallback_when_project_config_has_no_agent_field() {
-        let reg = registry_with("sonnet", &["sonnet"]);
-        let mut server = HarnessConfig::default();
-        server.agents.default_agent = "sonnet".to_string();
-
-        let dir = tempfile::tempdir().unwrap();
-        let harness_dir = dir.path().join(".harness");
-        std::fs::create_dir_all(&harness_dir).unwrap();
-        // Config file exists but has no [agent] section
-        std::fs::write(
-            harness_dir.join("config.toml"),
-            "[git]\nbase_branch = \"main\"\n",
-        )
-        .unwrap();
-
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-        let agent = select_agent(&req, &reg, None).unwrap();
-        assert_eq!(agent.name(), "sonnet");
-    }
-
-    // Regression for issue #2: a project that explicitly pins the same agent
-    // name as the server default must still honor tier 2 (i.e. bypass tier 3
-    // complexity dispatch). Previously the textual equality check
-    // `if &resolved != server_default` caused tier 2 to be silently skipped,
-    // allowing complex tasks to be rerouted by the complexity router.
-    #[test]
-    fn project_pinning_same_agent_as_server_default_bypasses_complexity_dispatch() {
-        // Registry has two agents; complexity dispatch would pick "opus" for
-        // complex tasks, but the project explicitly pins "claude" (same as
-        // server default) — tier 2 must win.
-        let reg = registry_with("claude", &["claude", "opus"]);
-        let mut server = HarnessConfig::default();
-        server.agents.default_agent = "claude".to_string();
-
-        let dir = tempfile::tempdir().unwrap();
-        let harness_dir = dir.path().join(".harness");
-        std::fs::create_dir_all(&harness_dir).unwrap();
-        // Explicitly pin the same agent as the server default.
-        std::fs::write(
-            harness_dir.join("config.toml"),
-            "[agent]\ndefault = \"claude\"\n",
-        )
-        .unwrap();
-
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-        let agent = select_agent(&req, &reg, None).unwrap();
-        // Must be "claude" (explicit project pin), not whatever complexity
-        // dispatch would have chosen.
-        assert_eq!(agent.name(), "claude");
-    }
-
-    #[test]
-    fn unknown_agent_in_project_config_returns_bad_request() {
-        let reg = registry_with("auto", &["sonnet"]);
-        let mut server = HarnessConfig::default();
-        server.agents.default_agent = "auto".to_string();
-
-        let dir = tempfile::tempdir().unwrap();
-        let harness_dir = dir.path().join(".harness");
-        std::fs::create_dir_all(&harness_dir).unwrap();
-        std::fs::write(
-            harness_dir.join("config.toml"),
-            "[agent]\ndefault = \"nonexistent\"\n",
-        )
-        .unwrap();
-
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-        let result = select_agent(&req, &reg, None);
-        assert!(matches!(result, Err(EnqueueTaskError::BadRequest(_))));
-    }
-
-    // Regression for Issue 1: `[agent] default = "auto"` in .harness/config.toml
-    // must NOT be treated as a literal agent name (which would 400). It is a
-    // sentinel meaning "fall through to complexity dispatch (tier 3)".
-    #[test]
-    fn config_toml_auto_falls_through_to_complexity_dispatch() {
-        let reg = registry_with("sonnet", &["sonnet"]);
-
-        let dir = tempfile::tempdir().unwrap();
-        let harness_dir = dir.path().join(".harness");
-        std::fs::create_dir_all(&harness_dir).unwrap();
-        std::fs::write(
-            harness_dir.join("config.toml"),
-            "[agent]\ndefault = \"auto\"\n",
-        )
-        .unwrap();
-
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-        // "auto" must not 400; it must fall through to complexity dispatch which
-        // returns the registry default ("sonnet").
-        let agent = select_agent(&req, &reg, None).unwrap();
-        assert_eq!(agent.name(), "sonnet");
-    }
-
-    // Regression for Issue 2: project registry default_agent (configured via
-    // POST /api/projects) must be honoured when .harness/config.toml has no
-    // [agent] section.
-    #[test]
-    fn registry_default_agent_used_when_config_toml_has_no_agent_section() {
-        let reg = registry_with("sonnet", &["sonnet", "opus"]);
-
-        // No .harness/config.toml → config.toml tier produces no agent.
-        let dir = tempfile::tempdir().unwrap();
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-
-        // Registry record says to use "opus".
-        let agent = select_agent(&req, &reg, Some("opus")).unwrap();
-        assert_eq!(agent.name(), "opus");
-    }
-
-    // "auto" in registry default_agent must also fall through to complexity
-    // dispatch (tier 3), not 400.
-    #[test]
-    fn registry_auto_agent_falls_through_to_complexity_dispatch() {
-        let reg = registry_with("sonnet", &["sonnet"]);
-
-        let dir = tempfile::tempdir().unwrap();
-        let req = req_with_agent(None, Some(dir.path().to_path_buf()));
-
-        let agent = select_agent(&req, &reg, Some("auto")).unwrap();
-        assert_eq!(agent.name(), "sonnet");
-    }
-}
+#[path = "task_routes_tests.rs"]
+mod tests;
